refactor(std): move text_encoding + promise_slot to rts-std (break async cycle)
O prereq do grupo async, sem arrastar promise/generator/fetch/blob/crypto junto.

Problema: text_encoding (microtask queue) move pro rts-std, mas drain_microtasks chama generator::async_sm_resume — pub fn no collector do rts-runtime. rts-std não pode depender de rts-runtime (ciclo).

Solução: wrapper extern. Adicionado __RTS_FN_RT_ASYNC_SM_RESUME(h,value,rejected:i64) #[no_mangle] no collector/generator.rs (runtime); text_encoding (std) chama via extern-decl (resolve por link). Único call std→runtime resolvido sem mover generator.

Movidos:
- promise_slot.rs (collector) -> rts-std/src/promise_slot.rs. super::handles -> rts_engine::heap::handles. Re-export em collector/mod.rs (pub use rts_std::promise_slot) — consumers runtime (promise/generator/fetch/blob/crypto) seguem via gc::promise_slot. Retornos Arc<PromiseSlot> cross-crate OK (PromiseSlot é do rts-engine).
- text_encoding (global) -> rts-std/src/globals/text_encoding. Refs: gc::handles->engine; collections::map + function::ops::invoke_fn_ptr_with_registry -> rts_shared; regex -> rts_shared::regex; promise_slot -> crate::; async_rt fica crate::runtime.

Cargo rts-std += rts-shared (path). SEM ciclo: rts-shared depende só de rts-engine; rts-std->rts-shared é descida de camada.

Gate: build limpo; rts.exe test 1710/1710 (async/Promise/microtask intactos); JIT smoke (TextEncoder/Decoder + await + Promise.then + queueMicrotask order, tudo correto).

PRE-EXISTING (não regressão, confirmado por inspeção do código): AOT-compiled binaries não drenam o microtask queue — drain_microtasks só é chamado no pipeline JIT (pipeline.rs), não emitido no epílogo do binário AOT. Logo await-tail/.then/queueMicrotask não disparam em AOT (sync APIs como TextEncoder funcionam). Memória project_aot_no_microtask_drain.
