+++ crates/rts-std/src/globals/mod.rs
@@ -16,3 +16,4 @@ pub mod event_target;
 pub mod message_channel;
 pub mod abort;
 pub mod events;
+pub mod text_encoding;
