+++ crates/rts-std/src/lib.rs
@@ -20,4 +20,5 @@ pub mod tls;
 pub mod thread;
 pub mod http_server;
 pub mod events;
+pub mod promise_slot;
 pub mod globals;
